Split flattener into multi test suites
