+++ tests/unit/downloader/content-flattened.json
@@ -1,0 +1,21 @@
+{
+  "Project": {
+    "headers": [
+      "project.uuid",
+      "project.project_core.project_short_name",
+      "project.project_core.project_title",
+      "project.project_core.project_description"
+    ],
+    "values": [
+      {
+        "project.uuid": "uuid1",
+        "project.project_core.project_short_name": "label",
+        "project.project_core.project_title": "title",
+        "project.project_core.project_description": "desc"
+      }
+    ]
+  },
+  "Schemas": [
+    "https://schema.humancellatlas.org/type/project/14.2.0/project"
+  ]
+}
