+++ tests/unit/downloader/entities-flattened.json
@@ -1,0 +1,67 @@
+{
+  "Project": {
+    "headers": [
+      "project.uuid",
+      "project.project_core.project_short_name",
+      "project.project_core.project_title",
+      "project.project_core.project_description"
+    ],
+    "values": [
+      {
+        "project.uuid": "uuid1",
+        "project.project_core.project_short_name": "label",
+        "project.project_core.project_title": "title",
+        "project.project_core.project_description": "desc"
+      }
+    ]
+  },
+  "Project - Contributors": {
+    "headers": [
+      "project.contributors.name",
+      "project.contributors.email",
+      "project.contributors.institution",
+      "project.contributors.laboratory",
+      "project.contributors.country",
+      "project.contributors.corresponding_contributor",
+      "project.contributors.project_role.text",
+      "project.contributors.project_role.ontology",
+      "project.contributors.project_role.ontology_label"
+    ],
+    "values": [
+      {
+        "project.contributors.corresponding_contributor": "True",
+        "project.contributors.country": "USA",
+        "project.contributors.email": "alex.pollen@ucsf.edu",
+        "project.contributors.institution": "University of California, San Francisco (UCSF)",
+        "project.contributors.laboratory": "Department of Neurology",
+        "project.contributors.name": "Alex A,,Pollen",
+        "project.contributors.project_role.ontology": "EFO:0009741",
+        "project.contributors.project_role.ontology_label": "experimental scientist",
+        "project.contributors.project_role.text": "experimental scientist"
+      }
+    ]
+  },
+  "Donor organism": {
+    "headers": [
+      "donor_organism.uuid",
+      "donor_organism.biomaterial_core.biomaterial_id",
+      "donor_organism.biomaterial_core.biomaterial_description"
+    ],
+    "values": [
+      {
+        "donor_organism.uuid": "uuid2",
+        "donor_organism.biomaterial_core.biomaterial_id": "label",
+        "donor_organism.biomaterial_core.biomaterial_description": "desc"
+      },
+      {
+        "donor_organism.uuid": "uuid3",
+        "donor_organism.biomaterial_core.biomaterial_id": "label",
+        "donor_organism.biomaterial_core.biomaterial_description": "desc"
+      }
+    ]
+  },
+  "Schemas": [
+    "https://schema.humancellatlas.org/type/project/14.2.0/project",
+    "https://schema.humancellatlas.org/type/project/14.2.0/donor_organism"
+            ]
+        }
