+++ tests/unit/downloader/entities.json
@@ -1,0 +1,57 @@
+[
+  {
+    "content": {
+      "describedBy": "https://schema.humancellatlas.org/type/project/14.2.0/project",
+      "schema_type": "project",
+      "project_core": {
+        "project_short_name": "label",
+        "project_title": "title",
+        "project_description": "desc"
+      },
+      "contributors": [
+        {
+          "name": "Alex A,,Pollen",
+          "email": "alex.pollen@ucsf.edu",
+          "institution": "University of California, San Francisco (UCSF)",
+          "laboratory": "Department of Neurology",
+          "country": "USA",
+          "corresponding_contributor": true,
+          "project_role": {
+            "text": "experimental scientist",
+            "ontology": "EFO:0009741",
+            "ontology_label": "experimental scientist"
+          }
+        }
+      ]
+    },
+    "uuid": {
+      "uuid": "uuid1"
+    }
+  },
+  {
+    "content": {
+      "describedBy": "https://schema.humancellatlas.org/type/project/14.2.0/donor_organism",
+      "schema_type": "biomaterial",
+      "biomaterial_core": {
+        "biomaterial_id": "label",
+        "biomaterial_description": "desc"
+      }
+    },
+    "uuid": {
+      "uuid": "uuid2"
+    }
+  },
+  {
+    "content": {
+      "describedBy": "https://schema.humancellatlas.org/type/project/14.2.0/donor_organism",
+      "schema_type": "biomaterial",
+      "biomaterial_core": {
+        "biomaterial_id": "label",
+        "biomaterial_description": "desc"
+      }
+    },
+    "uuid": {
+      "uuid": "uuid3"
+    }
+  }
+]
