+++ tests/unit/downloader/test_flattener.py
@@ -1,4 +1,5 @@
 import json
+import os
 from unittest import TestCase
 
 from ingest.downloader.flattener import Flattener
@@ -6,45 +7,25 @@ from ingest.downloader.flattener import Flattener
 
 class FlattenerTest(TestCase):
     def setUp(self) -> None:
-        self.content = {
-            "describedBy": "https://schema.humancellatlas.org/type/project/14.2.0/project",
-            "schema_type": "project",
-            "project_core": {
-                "project_short_name": "label",
-                "project_title": "title",
-                "project_description": "desc"
-            }
-        }
+        self.script_dir = os.path.dirname(__file__)
+
+        with open(self.script_dir + '/content.json') as file:
+            self.content = json.load(file)
 
         self.uuid = {
             'uuid': 'uuid1'
         }
 
-        self.flattened_metadata_entity = {
-            'Project': {
-                'headers': ['project.uuid', 'project.project_core.project_short_name',
-                            'project.project_core.project_title', 'project.project_core.project_description'],
-                'values': [
-                    {
-                        'project.uuid': 'uuid1',
-                        'project.project_core.project_short_name': 'label',
-                        'project.project_core.project_title': 'title',
-                        'project.project_core.project_description': 'desc'
-                    }
-                ]
-            },
-            'Schemas': [
-                'https://schema.humancellatlas.org/type/project/14.2.0/project'
-            ]
-        }
+        with open(self.script_dir + '/content-flattened.json') as file:
+            self.flattened_metadata_entity = json.load(file)
 
     def test_flatten__has_no_modules(self):
         # given
-        self.metadata_entity = {
+        metadata_entity = {
             'content': self.content,
             'uuid': self.uuid
         }
-        entity_list = [self.metadata_entity]
+        entity_list = [metadata_entity]
 
         # when
         flattener = Flattener()
@@ -88,8 +69,8 @@ class FlattenerTest(TestCase):
 
     def test_flatten__has_project_modules(self):
         # given
-        self.content.update({"contributors": [
-            {
+        self.content.update({
+            "contributors": [{
                 "name": "Alex A,,Pollen",
                 "email": "alex.pollen@ucsf.edu",
                 "institution": "University of California, San Francisco (UCSF)",
@@ -101,14 +82,15 @@ class FlattenerTest(TestCase):
                     "ontology": "EFO:0009741",
                     "ontology_label": "experimental scientist"
                 }
-            }
-        ]})
-        self.metadata_entity = {
+            }]
+        })
+
+        metadata_entity = {
             'content': self.content,
             'uuid': self.uuid
         }
 
-        entity_list = [self.metadata_entity]
+        entity_list = [metadata_entity]
 
         # when
         flattener = Flattener()
@@ -127,58 +109,18 @@ class FlattenerTest(TestCase):
                     'project.contributors.project_role.ontology',
                     'project.contributors.project_role.ontology_label'
                 ],
-                'values': [
-                    {'project.contributors.corresponding_contributor': 'True',
-                     'project.contributors.country': 'USA',
-                     'project.contributors.email': 'alex.pollen@ucsf.edu',
-                     'project.contributors.institution': 'University of California, San Francisco (UCSF)',
-                     'project.contributors.laboratory': 'Department of Neurology',
-                     'project.contributors.name': 'Alex A,,Pollen',
-                     'project.contributors.project_role.ontology': 'EFO:0009741',
-                     'project.contributors.project_role.ontology_label': 'experimental scientist',
-                     'project.contributors.project_role.text': 'experimental scientist'}
-                ]
-            }
+                'values': [{
+                    'project.contributors.corresponding_contributor': 'True',
+                    'project.contributors.country': 'USA',
+                    'project.contributors.email': 'alex.pollen@ucsf.edu',
+                    'project.contributors.institution': 'University of California, San Francisco (UCSF)',
+                    'project.contributors.laboratory': 'Department of Neurology',
+                    'project.contributors.name': 'Alex A,,Pollen',
+                    'project.contributors.project_role.ontology': 'EFO:0009741',
+                    'project.contributors.project_role.ontology_label': 'experimental scientist',
+                    'project.contributors.project_role.text': 'experimental scientist'}
+                ]}
         })
-
-        # then
-        self.assertEqual(actual, self.flattened_metadata_entity)
-
-    def test_flatten__has_ontology_property_with_single_element(self):
-        # given
-        self.content.update(
-            {"organ_parts": [
-                {
-                    "ontology": "UBERON:0000376",
-                    "ontology_label": "hindlimb stylopod",
-                    "text": "hindlimb stylopod"
-                }
-            ]})
-
-        self.metadata_entity = {
-            'content': self.content,
-            'uuid': self.uuid
-        }
-
-        entity_list = [self.metadata_entity]
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        self.flattened_metadata_entity['Project']['values'][0].update({
-            'project.organ_parts.ontology': 'UBERON:0000376',
-            'project.organ_parts.ontology_label': 'hindlimb stylopod',
-            'project.organ_parts.text': 'hindlimb stylopod',
-
-        })
-        self.flattened_metadata_entity['Project']['headers'].extend(
-            [
-                'project.organ_parts.ontology',
-                'project.organ_parts.ontology_label',
-                'project.organ_parts.text'
-            ]
-        )
 
         # then
         self.assertEqual(actual, self.flattened_metadata_entity)
@@ -197,203 +139,42 @@ class FlattenerTest(TestCase):
             ]
         }
 
-        self.metadata_entity = {
-            'content': self.content,
+        metadata_entity = {
+            'content': content,
             'uuid': self.uuid
         }
 
-        entity_list = [self.metadata_entity]
+        entity_list = [metadata_entity]
 
         # when
         flattener = Flattener()
         actual = flattener.flatten(entity_list)
         flattened_metadata_entity = {
-            'Collection Protocol': {
-                'values': [{
-
-                }],
+            'Collection protocol': {
+                'values': [{}],
                 'headers': []
             }
         }
-        flattened_metadata_entity['Collection Protocol']['values'][0].update({
+        flattened_metadata_entity['Collection protocol']['values'][0].update({
             'collection_protocol.organ_parts.field_1': 'UBERON:0000376',
             'collection_protocol.organ_parts.field_2': 'hindlimb stylopod',
             'collection_protocol.organ_parts.field_3': 'hindlimb stylopod',
-
+            'collection_protocol.uuid': 'uuid1'
         })
-        flattened_metadata_entity['Collection Protocol']['headers'].extend(
+        flattened_metadata_entity['Collection protocol']['headers'].extend(
             [
+                'collection_protocol.uuid',
                 'collection_protocol.organ_parts.field_1',
                 'collection_protocol.organ_parts.field_2',
                 'collection_protocol.organ_parts.field_3'
             ]
         )
+        flattened_metadata_entity['Schemas'] = [
+            'https://schema.humancellatlas.org/type/project/14.2.0/collection_protocol'
+        ]
 
         # then
-        self.assertEqual(actual, self.flattened_metadata_entity)
-
-    def test_flatten__has_ontology_property_with_multiple_elements(self):
-        # given
-        self.content.update(
-            {'organ_parts': [
-                {
-                    'ontology': 'UBERON:0000376',
-                    'ontology_label': 'dummylabel1',
-                    'text': 'dummytext1'
-                },
-                {
-                    'ontology': 'UBERON:0002386',
-                    'ontology_label': 'dummylabel2',
-                    'text': 'dummytext2'
-                }
-            ]})
-
-        self.metadata_entity = {
-            'content': self.content,
-            'uuid': self.uuid
-        }
-
-        entity_list = [self.metadata_entity]
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        self.flattened_metadata_entity['Project']['values'][0].update({
-            'project.organ_parts.ontology': 'UBERON:0000376||UBERON:0002386',
-            'project.organ_parts.ontology_label': 'dummylabel1||dummylabel2',
-            'project.organ_parts.text': 'dummytext1||dummytext2',
-
-        })
-        self.flattened_metadata_entity['Project']['headers'].extend(
-            [
-                'project.organ_parts.ontology',
-                'project.organ_parts.ontology_label',
-                'project.organ_parts.text'
-            ]
-        )
-
-        # then
-        self.assertEqual(actual, self.flattened_metadata_entity)
-
-    def test_flatten__has_ontology_property_with_multiple_elements_but_inconsistent_columns(self):
-        # given
-        self.content.update(
-            {'diseases': [
-                {
-                    'ontology': 'UBERON:0000376',
-                    'ontology_label': 'dummylabel1',
-                    'text': 'dummytext1'
-                },
-                {
-                    'text': 'dummytext2'
-                }
-            ]})
-
-        self.metadata_entity = {
-            'content': self.content,
-            'uuid': self.uuid
-        }
-
-        entity_list = [self.metadata_entity]
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        self.flattened_metadata_entity['Project']['values'][0].update({
-            'project.diseases.ontology': 'UBERON:0000376',
-            'project.diseases.ontology_label': 'dummylabel1',
-            'project.diseases.text': 'dummytext1||dummytext2',
-
-        })
-        self.flattened_metadata_entity['Project']['headers'].extend(
-            [
-                'project.diseases.ontology',
-                'project.diseases.ontology_label',
-                'project.diseases.text'
-            ]
-        )
-
-        # then
-        self.assertEqual(actual, self.flattened_metadata_entity)
-
-    def test_flatten__has_ontology_property_with_multiple_elements_but_with_empty_ontology_values(self):
-        # given
-        self.content.update(
-            {'diseases': [
-                {
-                    'ontology': 'UBERON:0000376',
-                    'ontology_label': 'dummylabel1',
-                    'text': 'dummytext1'
-                },
-                {
-                    'ontology': '',
-                    'ontology_label': '',
-                    'text': 'dummytext2'
-                }
-            ]})
-
-        self.metadata_entity = {
-            'content': self.content,
-            'uuid': self.uuid
-        }
-
-        entity_list = [self.metadata_entity]
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        self.flattened_metadata_entity['Project']['values'][0].update({
-            'project.diseases.ontology': 'UBERON:0000376',
-            'project.diseases.ontology_label': 'dummylabel1',
-            'project.diseases.text': 'dummytext1||dummytext2',
-
-        })
-        self.flattened_metadata_entity['Project']['headers'].extend(
-            [
-                'project.diseases.ontology',
-                'project.diseases.ontology_label',
-                'project.diseases.text'
-            ]
-        )
-
-        # then
-        self.assertEqual(self.flattened_metadata_entity, actual)
-
-    def test_flatten__has_ontology_property_with_single_element_but_only_with_text_attr(self):
-        # given
-        self.content.update(
-            {'diseases': [
-                {
-                    'text': 'dummytext2'
-                }
-            ]})
-
-        self.metadata_entity = {
-            'content': self.content,
-            'uuid': self.uuid
-        }
-
-        entity_list = [self.metadata_entity]
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        self.flattened_metadata_entity['Project']['values'][0].update({
-            'project.diseases.text': 'dummytext2',
-
-        })
-        self.flattened_metadata_entity['Project']['headers'].extend(
-            [
-                'project.diseases.text'
-            ]
-        )
-
-        # then
-        self.assertEqual(actual, self.flattened_metadata_entity)
+        self.assertEqual(actual, flattened_metadata_entity)
 
     def test_flatten__has_boolean(self):
         # given
@@ -454,154 +235,6 @@ class FlattenerTest(TestCase):
             },
             'Schemas': [
                 'https://schema.humancellatlas.org/type/project/14.2.0/project'
-            ]
-        }
-
-        # then
-        self.assertEqual(actual, expected)
-
-    def test_flatten__project_metadata(self):
-        self.maxDiff = None
-
-        # given
-        with open('project-list.json') as file:
-            entity_list = json.load(file)
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        with open('project-list-flattened.json') as file:
-            expected = json.load(file)
-
-        # then
-        self.assertEqual(actual, expected)
-
-    def test_flatten__has_different_entities(self):
-        # given
-        entity_list = [{
-            'content': {
-                "describedBy": "https://schema.humancellatlas.org/type/project/14.2.0/project",
-                "schema_type": "project",
-                "project_core": {
-                    "project_short_name": "label",
-                    "project_title": "title",
-                    "project_description": "desc"
-                },
-                "contributors": [
-                    {
-                        "name": "Alex A,,Pollen",
-                        "email": "alex.pollen@ucsf.edu",
-                        "institution": "University of California, San Francisco (UCSF)",
-                        "laboratory": "Department of Neurology",
-                        "country": "USA",
-                        "corresponding_contributor": True,
-                        "project_role": {
-                            "text": "experimental scientist",
-                            "ontology": "EFO:0009741",
-                            "ontology_label": "experimental scientist"
-                        }
-                    }
-                ]
-            },
-            'uuid': {
-                'uuid': 'uuid1'
-            }
-        },
-            {
-                'content': {
-                    "describedBy": "https://schema.humancellatlas.org/type/project/14.2.0/donor_organism",
-                    "schema_type": "biomaterial",
-                    "biomaterial_core": {
-                        "biomaterial_id": "label",
-                        "biomaterial_description": "desc"
-                    }
-                },
-                'uuid': {
-                    'uuid': 'uuid2'
-                }
-            },
-            {
-                'content': {
-                    "describedBy": "https://schema.humancellatlas.org/type/project/14.2.0/donor_organism",
-                    "schema_type": "biomaterial",
-                    "biomaterial_core": {
-                        "biomaterial_id": "label",
-                        "biomaterial_description": "desc"
-                    }
-                },
-                'uuid': {
-                    'uuid': 'uuid3'
-                }
-            }
-        ]
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        expected = {
-            'Project': {
-                'headers': [
-                    'project.uuid',
-                    'project.project_core.project_short_name',
-                    'project.project_core.project_title',
-                    'project.project_core.project_description'
-                ],
-                'values': [
-                    {
-                        'project.uuid': 'uuid1',
-                        'project.project_core.project_short_name': 'label',
-                        'project.project_core.project_title': 'title',
-                        'project.project_core.project_description': 'desc'
-                    }
-                ]},
-            'Project - Contributors': {
-                'headers': [
-                    'project.contributors.name',
-                    'project.contributors.email',
-                    'project.contributors.institution',
-                    'project.contributors.laboratory',
-                    'project.contributors.country',
-                    'project.contributors.corresponding_contributor',
-                    'project.contributors.project_role.text',
-                    'project.contributors.project_role.ontology',
-                    'project.contributors.project_role.ontology_label'
-                ],
-                'values': [
-                    {'project.contributors.corresponding_contributor': 'True',
-                     'project.contributors.country': 'USA',
-                     'project.contributors.email': 'alex.pollen@ucsf.edu',
-                     'project.contributors.institution': 'University of California, San Francisco (UCSF)',
-                     'project.contributors.laboratory': 'Department of Neurology',
-                     'project.contributors.name': 'Alex A,,Pollen',
-                     'project.contributors.project_role.ontology': 'EFO:0009741',
-                     'project.contributors.project_role.ontology_label': 'experimental scientist',
-                     'project.contributors.project_role.text': 'experimental scientist'}
-                ]
-            },
-            'Donor organism': {
-                'headers': [
-                    'donor_organism.uuid',
-                    'donor_organism.biomaterial_core.biomaterial_id',
-                    'donor_organism.biomaterial_core.biomaterial_description'
-                ],
-                'values': [
-                    {
-                        'donor_organism.uuid': 'uuid2',
-                        'donor_organism.biomaterial_core.biomaterial_id': 'label',
-                        'donor_organism.biomaterial_core.biomaterial_description': 'desc'
-                    },
-                    {
-                        'donor_organism.uuid': 'uuid3',
-                        'donor_organism.biomaterial_core.biomaterial_id': 'label',
-                        'donor_organism.biomaterial_core.biomaterial_description': 'desc'
-                    }
-                ]
-            },
-            'Schemas': [
-                'https://schema.humancellatlas.org/type/project/14.2.0/project',
-                'https://schema.humancellatlas.org/type/project/14.2.0/donor_organism'
             ]
         }
 
@@ -673,21 +306,6 @@ class FlattenerTest(TestCase):
         # then
         self.assertEqual(actual, expected)
 
-    def test_flatten__collection_protocol_metadata(self):
-        # given
-        with open('collection_protocol-list.json') as file:
-            entity_list = json.load(file)
-
-        # when
-        flattener = Flattener()
-        actual = flattener.flatten(entity_list)
-
-        with open('collection_protocol-list-flattened.json') as file:
-            expected = json.load(file)
-
-        # then
-        self.assertEqual(actual, expected)
-
     def test_flatten__raises_error__given_multiple_schema_versions_of_same_concrete_entity(self):
         # given
         entity_list = [
@@ -718,4 +336,3 @@ class FlattenerTest(TestCase):
         flattener = Flattener()
         with self.assertRaisesRegex(ValueError, "Multiple versions of same concrete entity schema"):
             flattener.flatten(entity_list)
-
